requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
flask>=3.0.0
reportlab